+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 
 - Resolved a "Slow operations are prohibited on EDT" exception on Flutter Project creation (#8446, #8447, #8448)
 - Made dev release daily instead of weekly
+- Set the device selector component to opaque during its creation to avoid an unexpected background color (#8471)
 
 ## 87.1.0